app/requirements.txt
fastapi[standard]
pydantic
redis
influxdb-client
pandas
python-dateutil
